[process_modules] restructured process_modules and cleaned up a bit
