+++ src/pycram/__init__.py
@@ -19,6 +19,7 @@ process_module -- implementation of process modules.
 """
 import logging
 import logging.config
+import pycram.process_modules.available_process_modules
 
 logging.basicConfig(level=logging.WARNING, format='%(levelname)s - %(name)s  - Line:%(lineno)d - %(message)s')
