+++ src/pycram/process_modules/available_process_modules.py
@@ -1,9 +1,9 @@
-from pycram.robot_description import InitializedRobotDescription as robot_description
-from pycram.process_module import ProcessModule
-from pr2_process_modules import PR2ProcessModulesSimulated, PR2ProcessModulesReal
-from boxy_process_modules import BoxyProcessModulesSimulated, BoxyProcessModulesReal
-from donbot_process_modules import DonbotProcessModulesSimulated, DonbotProcessModulesReal
-from hsr_process_modules import HSRProcessModulesSimulated, HSRProcessModulesReal
+from ..robot_description import InitializedRobotDescription as robot_description
+from ..process_module import ProcessModule
+from .pr2_process_modules import PR2ProcessModulesSimulated, PR2ProcessModulesReal
+from .boxy_process_modules import BoxyProcessModulesSimulated, BoxyProcessModulesReal
+from .donbot_process_modules import DonbotProcessModulesSimulated, DonbotProcessModulesReal
+from .hsr_process_modules import HSRProcessModulesSimulated, HSRProcessModulesReal
 from rospy import logerr
