+++ src/pycram/process_modules/boxy_process_modules.py
@@ -1,19 +1,19 @@
-from pycram.robot_description import InitializedRobotDescription as robot_description
-from pycram.process_module import ProcessModule
-from pycram.bullet_world import BulletWorld
-from pycram.helper import transform
-from pycram.ik import request_ik
-from pycram.helper import _transform_to_torso, _apply_ik
-from pycram.local_transformer import local_transformer
+import pycram.helper_deprecated as helper_deprecated
+from ..robot_description import InitializedRobotDescription as robot_description
+from ..process_module import ProcessModule
+from ..bullet_world import BulletWorld
+from ..ik import request_ik
+from ..local_transformer import local_transformer as local_tf
+import pycram.helper as helper
 import pycram.bullet_world_reasoning as btr
+
 import pybullet as p
-import numpy as np
 import time
 
 
 def _park_arms(arm):
     """
-    Defines the joint poses for the parking positions of the arms of PR2 and applies them to the
+    Defines the joint poses for the parking positions of the arms of Boxy and applies them to the
     in the BulletWorld defined robot.
     :return: None
     """
@@ -27,7 +27,7 @@ def _park_arms(arm):
             robot.set_joint_state(joint, pose)
 
 
-class Pr2Navigation(ProcessModule):
+class BoxyNavigation(ProcessModule):
     """
     The process module to move the robot from one position to another.
     """
@@ -36,12 +36,17 @@ class Pr2Navigation(ProcessModule):
         solution = desig.reference()
         if solution['cmd'] == 'navigate':
             robot = BulletWorld.robot
+            # Reset odom joints to zero
+            for joint_name in robot_description.i.odom_joints:
+                robot.set_joint_state(joint_name, 0.0)
+            # Set actual goal pose
             robot.set_position_and_orientation(solution['target'], solution['orientation'])
             time.sleep(0.5)
-            local_transformer.update_from_btr()
+            local_tf.update_from_btr()
 
 
-class Pr2PickUp(ProcessModule):
+
+class BoxyPickUp(ProcessModule):
     """
     This process module is for picking up a given object.
     The object has to be reachable for this process module to succeed.
@@ -49,23 +54,24 @@ class Pr2PickUp(ProcessModule):
 
     def _execute(self, desig):
         solution = desig.reference()
-        if solution['cmd'] == 'pick-up':
+        if solution['cmd'] == 'pick':
             object = solution['object']
             robot = BulletWorld.robot
-            #target = object.get_position_and_orientation()
             grasp = robot_description.i.grasps.get_orientation_for_grasp(solution['grasp'])
             target = [object.get_position(), grasp]
-            target = _transform_to_torso(target, robot)
+            target = helper._transform_to_torso(target, robot)
             arm = "left" if solution['gripper'] == robot_description.i.get_tool_frame("left") else "right"
             joints = robot_description.i._safely_access_chains(arm).joints
             #tip = "r_wrist_roll_link" if solution['gripper'] == "r_gripper_tool_frame" else "l_wrist_roll_link"
             inv = request_ik(robot_description.i.base_frame, solution['gripper'], target, robot, joints)
-            _apply_ik(robot, inv, solution['gripper'])
+            helper._apply_ik(robot, inv, solution['gripper'])
+            #inv = p.calculateInverseKinematics(robot.id, robot.get_link_id(solution['gripper']), target,
+            #                                   maxNumIterations=100)
             robot.attach(object, solution['gripper'])
             time.sleep(0.5)
 
 
-class Pr2Place(ProcessModule):
+class BoxyPlace(ProcessModule):
     """
     This process module places an object at the given position in world coordinate frame.
     """
@@ -75,21 +81,18 @@ class Pr2Place(ProcessModule):
         if solution['cmd'] == 'place':
             object = solution['object']
             robot = BulletWorld.robot
-            target = [solution['target'], [0, 0, 0, 1]]
-            target = _transform_to_torso(target, robot)
-            target = (target[0], [0, 0, 0, 1])
+            target = object.get_position_and_orientation()
+            target = helper._transform_to_torso(target, robot)
             arm = "left" if solution['gripper'] == robot_description.i.get_tool_frame("left") else "right"
             joints = robot_description.i._safely_access_chains(arm).joints
-            #inv = p.calculateInverseKinematics(robot.id, robot.get_link_id(solution['gripper']), solution['target'],
-            #                               maxNumIterations=100)
             #tip = "r_wrist_roll_link" if solution['gripper'] == "r_gripper_tool_frame" else "l_wrist_roll_link"
             inv = request_ik(robot_description.i.base_frame, solution['gripper'], target, robot, joints)
-            _apply_ik(robot, inv, solution['gripper'])
+            helper._apply_ik(robot, inv, solution['gripper'])
             robot.detach(object)
             time.sleep(0.5)
 
 
-class Pr2Accessing(ProcessModule):
+class BoxyAccessing(ProcessModule):
     """
     This process module responsible for opening drawers to access the objects inside. This works by firstly moving
     the end effector to the handle of the drawer. Next, the end effector is moved the respective distance to the back.
@@ -106,23 +109,27 @@ class Pr2Accessing(ProcessModule):
             drawer_handle = solution['drawer_handle']
             drawer_joint = solution['drawer_joint']
             dis = solution['distance']
+            robot.set_joint_state(robot_description.i.torso_joint, -0.1)
             arm = "left" if solution['gripper'] == robot_description.i.get_tool_frame("left") else "right"
             joints = robot_description.i._safely_access_chains(arm).joints
             #inv = p.calculateInverseKinematics(robot.id, robot.get_link_id(gripper), kitchen.get_link_position(drawer_handle))
-            target = _transform_to_torso(kitchen.get_link_position_and_orientation(drawer_handle), robot)
-            target = (target[0], [0, 0, 0, 1])
+            target = helper._transform_to_torso(kitchen.get_link_position_and_orientation(drawer_handle), robot)
+            #target = (target[0], [0, 0, 0, 1])
             inv = request_ik(robot_description.i.base_frame, gripper, target , robot, joints )
-            _apply_ik(robot, inv, gripper)
+            helper._apply_ik(robot, inv, gripper)
             time.sleep(0.2)
-            new_p = ([target[0][0] - dis, target[0][1], target[0][2]],target[1])
-            #inv = p.calculateInverseKinematics(robot.id, robot.get_link_id(gripper), new_p)
+            cur_pose = robot.get_pose()
+            robot.set_position([cur_pose[0]-dis, cur_pose[1], cur_pose[2]])
+            han_pose = kitchen.get_link_position(drawer_handle)
+            new_p = [[han_pose[0] - dis, han_pose[1], han_pose[2]], kitchen.get_link_orientation(drawer_handle)]
+            new_p = helper._transform_to_torso(new_p, robot)
             inv = request_ik(robot_description.i.base_frame, gripper, new_p, robot, joints)
-            _apply_ik(robot, inv, gripper)
-            kitchen.set_joint_state(drawer_joint, dis)
+            helper._apply_ik(robot, inv, gripper)
+            kitchen.set_joint_state(drawer_joint, 0.3)
             time.sleep(0.5)
 
 
-class Pr2ParkArms(ProcessModule):
+class BoxyParkArms(ProcessModule):
     """
     This process module is for moving the arms in a parking position.
     It is currently not used.
@@ -134,7 +141,8 @@ class Pr2ParkArms(ProcessModule):
             _park_arms()
 
 
-class Pr2MoveHead(ProcessModule):
+
+class BoxyMoveHead(ProcessModule):
     """
     This process module moves the head to look at a specific point in the world coordinate frame.
     This point can either be a position or an object.
@@ -143,24 +151,48 @@ class Pr2MoveHead(ProcessModule):
     def _execute(self, desig):
         solutions = desig.reference()
         if solutions['cmd'] == 'looking':
-            target = solutions['target']
             robot = BulletWorld.robot
-            if type(target) is str:
-                target_frame = local_transformer.projection_namespace + '/' + target \
-                    if local_transformer.projection_namespace \
-                    else target
-                target = local_transformer.tf_transform(local_transformer.map_frame, target_frame)[0]
-            pose_in_pan = transform(target, robot.get_link_position("head_pan_link"))
-            pose_in_tilt = transform(target, robot.get_link_position("head_tilt_link"))
+            neck_base_frame = local_tf.projection_namespace + '/' + robot_description.i.chains["neck"].base_link
+            if type(solutions['target']) is str:
+                target = local_tf.projection_namespace + '/' + solutions['target']
+                pose_in_neck_base = local_tf.tf_transform(neck_base_frame, target)
+            elif helper_deprecated.is_list_pose(solutions['target']) or helper_deprecated.is_list_position(solutions['target']):
+                pose = helper_deprecated.ensure_pose(solutions['target'])
+                pose_in_neck_base = local_tf.tf_pose_transform(local_tf.map_frame, neck_base_frame, pose)
 
-            new_pan = np.arctan([pose_in_pan[1], pose_in_pan[0]])
-            new_tilt = np.arctan([-pose_in_tilt[2], pose_in_tilt[0] ** 2 + pose_in_tilt[1] ** 2])
+            vector = pose_in_neck_base[0]
+            # +x as forward
+            # +y as left
+            # +z as up
+            x = vector[1]
+            y = -vector[0]
+            z = vector[2]
+            conf = None
+            if x > 0:
+                if z < 0.5:
+                    if y > 0.4:
+                        conf = "down_left"
+                    elif y < -0.4:
+                        conf = "down_right"
+                    else:
+                        conf = "down"
+                else:
+                    if y > 0.4:
+                        conf = "left"
+                    elif y < -0.4:
+                        conf = "right"
+                    else:
+                        conf = "forward"
+            else:
+                if z < 0.5:
+                    conf = "behind"
+                else:
+                    conf = "behind_up"
+            for joint, state in robot_description.i.get_static_joint_chain("neck", conf).items():
+                robot.set_joint_state(joint, state)
 
-            robot.set_joint_state("head_pan_joint", new_pan[0])
-            robot.set_joint_state("head_tilt_joint", new_tilt[0])
 
-
-class Pr2MoveGripper(ProcessModule):
+class BoxyMoveGripper(ProcessModule):
     """
     This process module controls the gripper of the robot. They can either be opened or closed.
     Furthermore, it can only moved one gripper at a time.
@@ -173,11 +205,12 @@ class Pr2MoveGripper(ProcessModule):
             gripper = solution['gripper']
             motion = solution['motion']
             for joint, state in robot_description.i.get_static_gripper_chain(gripper, motion).items():
+                # TODO: Test this, add gripper-opening/-closing to the demo.py
                 robot.set_joint_state(joint, state)
             time.sleep(0.5)
 
 
-class Pr2Detecting(ProcessModule):
+class BoxyDetecting(ProcessModule):
     """
     This process module tries to detect an object with the given type. To be detected the object has to be in
     the field of view of the robot.
@@ -193,11 +226,11 @@ class Pr2Detecting(ProcessModule):
 
             objects = BulletWorld.current_bullet_world.get_objects_by_type(object_type)
             for obj in objects:
-                if btr.visible(obj, robot.get_link_position_and_orientation(cam_frame_name), front_facing_axis):
+                if btr.visible(obj, robot.get_link_position_and_orientation(cam_frame_name), front_facing_axis, 0.5):
                     return obj
 
 
-class Pr2MoveTCP(ProcessModule):
+class BoxyMoveTCP(ProcessModule):
     """
     This process moves the tool center point of either the right or the left arm.
     """
@@ -206,17 +239,14 @@ class Pr2MoveTCP(ProcessModule):
         solution = desig.reference()
         if solution['cmd'] == "move-tcp":
             target = solution['target']
-            target = _transform_to_torso([target, [0, 0, 0, 1]])
             gripper = solution['gripper']
             robot = BulletWorld.robot
-            joints = ik_joints_left if gripper == "l_gripper_tool_frame" else ik_joints_right
-            #inv = p.calculateInverseKinematics(robot.id, robot.get_link_id(gripper), target)
-            inv = request_ik(pr2_root_link, gripper, target, robot, joints)
-            _apply_ik(robot, inv, gripper)
+            inv = p.calculateInverseKinematics(robot.id, robot.get_link_id(gripper), target)
+            helper._apply_ik(robot, inv)
             time.sleep(0.5)
 
 
-class Pr2MoveJoints(ProcessModule):
+class BoxyMoveJoints(ProcessModule):
     """
     This process modules moves the joints of either the right or the left arm. The joint states can be given as
     list that should be applied or a pre-defined position can be used, such as "parking"
@@ -228,7 +258,6 @@ class Pr2MoveJoints(ProcessModule):
             robot = BulletWorld.robot
             right_arm_poses = solution['right_arm_poses']
             left_arm_poses = solution['left_arm_poses']
-
             if type(right_arm_poses) == dict:
                 for joint, pose in right_arm_poses.items():
                     robot.set_joint_state(joint, pose)
@@ -244,7 +273,7 @@ class Pr2MoveJoints(ProcessModule):
             time.sleep(0.5)
 
 
-class Pr2WorldStateDetecting(ProcessModule):
+class BoxyWorldStateDetecting(ProcessModule):
     """
     This process module detectes an object even if it is not in the field of view of the robot.
     """
@@ -255,16 +284,17 @@ class Pr2WorldStateDetecting(ProcessModule):
             obj_type = solution['object_type']
             return list(filter(lambda obj: obj.type == obj_type, BulletWorld.current_bullet_world.objects))[0]
 
-PR2ProcessModulesSimulated = {'navigate' : Pr2Navigation(),
-                              'pick-up' : Pr2PickUp(),
-                              'place' : Pr2Place(),
-                              'access' : Pr2Accessing(),
-                              'looking' : Pr2MoveHead(),
-                              'opening_gripper' : Pr2MoveGripper(),
-                              'closing_gripper' : Pr2MoveGripper(),
-                              'detecting' : Pr2Detecting(),
-                              'move-tcp' : Pr2MoveTCP(),
-                              'move-joints' : Pr2MoveJoints(),
-                              'world-state-detecting' : Pr2WorldStateDetecting()}
 
-PR2ProcessModulesReal = {}
+BoxyProcessModulesSimulated = {'moving' : BoxyNavigation(),
+                              'pick-up' : BoxyPickUp(),
+                              'place' : BoxyPlace(),
+                              'accessing' : BoxyAccessing(),
+                              'looking' : BoxyMoveHead(),
+                              'opening_gripper' : BoxyMoveGripper(),
+                              'closing_gripper' : BoxyMoveGripper(),
+                              'detecting' : BoxyDetecting(),
+                              'move-tcp' : BoxyMoveTCP(),
+                              'move-arm-joints' : BoxyMoveJoints(),
+                              'world-state-detecting' : BoxyWorldStateDetecting()}
+
+BoxyProcessModulesReal = {}
